[fontc] Add workload module
This is no functional change, although I did move some of the code
related to testing into a method, Workload::run_for_test; this saved me
needing to make a bunch of the fields on Workload visible.
